requirements.txt
ProfanityBlocker
discord.py